Geometry

SmoothedParticles.Ball — Type

Ball(x1::Float64, x2::Float64, x3::Float64, r::Float64) <: Shape

Define a ball by specifying the center (x1, x2, x3) and the radius r.

SmoothedParticles.BooleanDifference — Type

BooleanDifference(s1::Shape, s2::Shape) <: Shape

Define a shape as difference of two shapes. Equivalent to s1 - s2.

SmoothedParticles.BooleanIntersection — Type

BooleanIntersection(s1::Shape, s2::Shape) <: Shape

Define a shape as intersection of two shapes. Equivalent to s1 * s2.

SmoothedParticles.BoundaryLayer — Type

BoundaryLayer(s::Shape, grid::Grid, width::Float64) <: Shape

Creates a layer of certain width around shape s. More specifically, a point is inside boundary layer if it is not in s and at the same time has distance less than width to at least one point on grid in s.

SmoothedParticles.Box — Type

Box(x1_min::Float64, x2_min::Float64, x3_min::Float64,
    x1_max::Float64, x2_max::Float64, x3_max::Float64)::Shape

Defines a box specified by two corners.

SmoothedParticles.Circle — Type

Circle(x1::Float64, x2::Float64, r::Float64) <: Shape

Define a circle by specifying the center (x1, x2) and the radius r.

SmoothedParticles.Ellipse — Type

Ellipse(x1::Float64, x2::Float64, r1::Float64, r2::Float64) <: Shape

Define an ellipse by specifying the center (x1, x2) and semi-major/minor axes r1, r2.

SmoothedParticles.Specification — Type

Specification(s::Shape, f::Function) <: Shape

Define a shape of all x in s, such that f(x) == true.

SmoothedParticles.Transform — Type

Transform(s::Shape; A::RealMatrix = MAT1, b::RealVector = VEC0)

Define shape as a linear transform $x \to Ax + b$ applied to shape s.

SmoothedParticles.Rectangle — Method

Rectangle(x1_min, x2_min, x1_max, x2_max) <: Shape

Define a rectangle by specifying bottom left and top right corner.