Kernels

Smoothing kernels are the guts of SPH. They measure the strength of interaction between neighbouring particles based on their distance. They are also used to interpolate particle variables into continuous Eulerian fields.

SmoothedParticles.Dspline23 — Method

Dspline23(h::Float64, r::Float64)::Float64

Returns $\frac{\text{d}w}{\text{d}r}(r)$, the first derivative of a 2d cubic spline $w$ with support radius h. Integrates to unity.

SmoothedParticles.Dspline24 — Method

Dspline24(h::Float64, r::Float64)::Float64

Returns $\frac{\text{d}w}{\text{d}r}(r)$, the first derivative of a 2d quartic spline $w$ with support radius h. Integrates to unity.

SmoothedParticles.Dwendland2 — Method

Dwendland2(h::Float64, r::Float64)::Float64

Returns $\frac{\text{d}w}{\text{d}r}(r)$, the first derivative of a 2d quintic Wendland kernel $w$ with support radius h. Integrates to unity.

SmoothedParticles.Dwendland3 — Method

Dwendland3(h::Float64, r::Float64)::Float64

Returns $\frac{\text{d}w}{\text{d}r}(r)$, the first derivative of a 3d quintic Wendland kernel $w$ with support radius h. Integrates to unity.

SmoothedParticles.rDspline23 — Method

rDspline23(h::Float64, r::Float64)::Float64

Returns $\frac{1}{r}\,\frac{\text{d}w}{\text{d}r}(r)$, the reduced first derivative of a 2d cubic spline $w$ with support radius h. Integrates to unity.

SmoothedParticles.rDspline24 — Method

rDspline24(h::Float64, r::Float64)::Float64

Returns $\frac{1}{r}\,\frac{\text{d}w}{\text{d}r}(r)$, the reduced first derivative of a 2d quartic spline $w$ with support radius h. Integrates to unity.

SmoothedParticles.rDwendland2 — Method

rDwendland2(h::Float64, r::Float64)::Float64

Returns $\frac{1}{r}\,\frac{\text{d}w}{\text{d}r}(r)$, the reduced first derivative of a 2d quintic Wendland kernel $w$ with support radius h. Integrates to unity.

SmoothedParticles.rDwendland3 — Method

rDwendland3(h::Float64, r::Float64)::Float64

Returns $\frac{1}{r}\,\frac{\text{d}w}{\text{d}r}(r)$, the reduced first derivative of a 3d quintic Wendland kernel $w$ with support radius h. Integrates to unity.

SmoothedParticles.spline23 — Method

spline23(h::Float64, r::Float64)::Float64

Returns $w(r)$, the value of a 2d cubic spline $w$ with support radius h. Integrates to unity.

SmoothedParticles.spline24 — Method

spline24(h::Float64, r::Float64)::Float64

Returns $w(r)$, the value of a 2d quartic spline $w$ with support radius h. Integrates to unity.

SmoothedParticles.wendland2 — Method

wendland2(h::Float64, r::Float64)::Float64

Returns $w(r)$, the value of a 2d quintic Wendland kernel $w$ with support radius h. Integrates to unity.

SmoothedParticles.wendland3 — Method

wendland3(h::Float64, r::Float64)::Float64

Returns $w(r)$, the value of a 3d quintic Wendland kernel $w$ with support radius h. Integrates to unity.