Geometry

Main.SPHLib.default_symmetry — Constant

default_symmetry = "hexagonal"

Main.SPHLib.BooleanDifference — Type

BooleanDifference(s1::Shape, s2::Shape) <: Shape

Define a shape as difference of two shapes. Equivalent to s1 - s2.

Main.SPHLib.BooleanIntersection — Type

BooleanIntersection(s1::Shape, s2::Shape) <: Shape

Define a shape as intersection of two shapes. Equivalent to s1 * s2.

Main.SPHLib.BoundaryLayer — Type

BoundaryLayer(s::Shape, dr::Float64, width::Float64; symmetry = default_symmetry) <: Shape

Creates a layer of certain width around shape s. This requires some details about the discretization, namely dr (characterstic length) and symmetry.

Supported values of symmetry are "hexagonal" or "square".

Main.SPHLib.Circle — Type

Circle(x::Float64, y::Float64, r::Float64) <: Shape

Define a circle by specifying the center (x, y) and the radius r.

Main.SPHLib.Ellipse — Type

Ellipse(x::Float64, y::Float64, r1::Float64, r2::Float64) <: Shape

Define an ellipse by specifying the center (x, y) and two semi-major/minor axes r1, r2.

Main.SPHLib.Polygon — Type

Polygon(v::Tuple{Float64, Float64}...) <: Shape

Define a polygon by specifying all vortices.

Main.SPHLib.Rectangle — Type

Rectangle(xrange::Tuple{Float64, Float64}, yrange::Tuple{Float64, Float64}) <: Shape

Define a rectangle as a cartesian product of two intervals.

Main.SPHLib.Shape — Type

Shape

Supertype for 2d geometrical shapes. Every user-defined subtype T <: Shape is expected to support method

is_inside(x::Float64, y::Float64, s::Shape) -> Bool

which determines whether a given point (x, y) lies inside s.

Main.SPHLib.Specification — Type

Specification(s::Shape, f::Function) <: Shape

Define a shape of all (x, y) in s, such that f(x,y) == true.

Main.SPHLib.generate_particles! — Method

generate_particles!(sys::ParticleSystem,
                    char_function::Function,
                    constructor::Function;
                    dr::Float64 = sys.dr,
                    symmetry::String = default_symmetry)

Create particles using constructor(x::Float64, y::Float64)::AbstractParticle at every point, where char_function(x::Float64, y::Float64)::Bool is true. The density of particles will be $\frac{1}{\text{d}r^2}$.

Supported values of symmetry are "hexagonal" or "square".

Main.SPHLib.generate_particles! — Method

generate_particles!(sys::ParticleSystem,
                    geometry::Shape,
                    constructor::Function;
                    dr::Float64 = sys.dr,
                    symmetry::String = default_symmetry)

Create particles using constructor(x::Float64, y::Float64)::AbstractParticle at every point inside a given shape. The density of particles will be $\frac{1}{\text{d}r^2}$.

Supported values of symmetry are "hexagonal" or "square".

Main.SPHLib.remove_particles! — Method

remove_particles!(sys::ParticleSystem, criterion::Function)

Remove all particles p in sys satisfying criterion(p) == true.

Main.SPHLib.snap_to_grid — Method

snap_to_grid(sys::ParticleSystem,
             x::Float64, y::Float64;
             symmetry::String = default_symmetry
             )::Tuple{Float64, Float64}

Find the nearest point to (x, y) which will be inside a grid.