Grids

SmoothedParticles.Grid — Method

Grid(dr::Float64, symm::Symbol)::Grid

Define a grid with a given characterstic length $\text{d}r$ and symmetry type symm. Characterstic length means that one particle occupies a volume $\text{d}r^n$, where $n$ is the dimension. Supported symmetry types are

2D:

:square (square grid)

:hexagonal (hexagrid, dual of isometric grid)

:vogel (grid based on Vogel's spiral, see: https://en.wikipedia.org/wiki/Fermat%27s_spiral)

3D:

:cubic (cubic grid, see: https://en.wikipedia.org/wiki/Cubiccrystalsystem)

:facecentered (cubic face-centered grid)

:bodycentered (cubic body-centered grid)

:diamond (diamond cubic grid, see: https://en.wikipedia.org/wiki/Diamond_cubic)

SmoothedParticles.generate_particles! — Method

generate_particles!(sys::ParticleSystem,
                    grid::Grid,
                    geometry::Shape,
                    constructor::Function)

Create particles using constructor(x::RealVector)::AbstractParticle at every grid point inside a given shape.