Structs

SmoothedParticles.AbstractParticle — Type

AbstractParticle

Abstract supertype for smoothed particles. Any structure with AbstractParticle supertype is expected to:

be mutable

have field x::RealVector (particle position)

SmoothedParticles.ParticleField — Type

ParticleField(sys::ParticleSystem, varS::Symbol)

Creates an abstract array whose $n$-th element is the value of scalar varS of $n$-th particle in sys.

Warning

The indentity of $n$-th particle in ParticleSystem may change when particles are added or removed.

SmoothedParticles.ParticleSystem — Type

ParticleSystem(T::Type, domain::Shape, h::Float64)

Struct that contains all vital information about the simulation. The constructor specifies that:

the simulation will use particles of type T <: AbstractParticle,

Particles outside of the 'domain' can be disregarded (and will be automatically removed).

Particles are considered neighbours if their distance is less than h.

Please, do not make 'domain' unnecessarily large (has negative impact on performance).

SmoothedParticles.Shape — Type

Shape

Supertype for geometrical shapes.