Structs

Main.SPHLib.AbstractParticle — Type

AbstractParticle

Abstract supertype for smoothed particles. Any structure with AbstractParticle supertype is expected to:

be mutable

have fields x::Float64 and y::Float64

Main.SPHLib.ParticleSystem — Type

ParticleSystem(T::Type;
               xlims::Tuple{Float64, Float64},
               ylims::Tuple{Float64, Float64},
               h::Float64)

An immutable struct that contains all vital information about the simulation. The constructor specifies that:

the simulation will use particles of type T <: AbstractParticle,

it will take place inside a cartesian product xlims times ylims,

Particles are considered neighbours if their distance is less than h.

Individual particles can be accessed through attribute ParticleSystem.particles

Main.SPHLib.ScalarField — Type

ScalarField(sys::ParticleSystem, varS::Symbol, name::String = string(varS))

Creates an abstract array whose $n$-th element is the value of scalar varS of $n$-th particle in sys. The name will be used when saved into a file.

Main.SPHLib.VectorField — Type

VectorField(sys::ParticleSystem, varS::Tuple{Symbol, Symbol}, name::String = string(varS))

Creates an abstract array whose $n$-th element is the value of vector varS of $n$-th particle in sys. The name will be used when saved into a file.

Main.SPHLib.@define_particle — Macro

@define_particle(name::Symbol, vars::Symbol...)

Macro for quick definition of particles. Defines a struct with given name that is a subtype of AbstractParticle and has fields x, y and other fields specified in vars. Comes equipped with a default constructor that takes x, y as positional and vars as optionl keyword arguments (default to 0.0)

Example:

@define_particle Particle Dx Dy rho

is equivalent to

mutable struct Particle <: AbstractParticle
	x::Float64
	y::Float64
	Dx::Float64
	Dy::Float64
	rho::Float64
	Particle(x, y; Dx = 0.0, Dy = 0.0, rho = 0.0) = new(x, y, Dx, Dy, rho)
end

Warning

The indentity of $n$-th particle in ParticleSystem may change when particles are added or removed.